Work on threads

+++ dto/src/dto/detail/DTOTermination.java
@@ -3,9 +3,9 @@ package dto.detail;
 public class DTOTermination {
 
     private final Long seconds;
-    private final Integer ticks;
+    private final Long ticks;
 
-    public DTOTermination(Long seconds, Integer ticks) {
+    public DTOTermination(Long seconds, Long ticks) {
         this.seconds = seconds;
         this.ticks = ticks;
     }
@@ -14,7 +14,7 @@ public class DTOTermination {
         return seconds;
     }
 
-    public Integer getTicks() {
+    public Long getTicks() {
         return ticks;
     }
 }
